Fix treatment of dots at start of line
Periods at the start of a line cause issues for groff.  The normal 'escape'
sequence for periods (\.) causes lines to disappear when they are the first
non-whitespace characters, even though the escape works elsewhere.

One solution is to write periods at the start of a line as \[char46] (see
http://stackoverflow.com/questions/11469341/escaping-dots-in-groff).  This
commit adds this special escape and adds-to or fixes tests as appropriate.

Conflicts:
	test/dots_at_line_start_test.roff
	test/dots_at_line_start_test.ronn